backend/requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
gunicorn==20.0.4
python-dotenv==0.10.3
django-heroku==0.3.1
django-filter==2.2.0